requirements.txt
pandas==1.2.4
mysql-connector-python==8.0.28
python-decouple==3.5
urllib3==1.25.11